requirements.txt
amqp==2.6.1
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.7
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.11.0
drf-yasg==1.20.0
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==1.1.1
numpy==1.19.1
packaging==20.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pylibmc==1.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.0.1
xlrd==1.2.0
zipp==3.0.0
django-elasticsearch-dsl==7.2.2
django-nine==0.2.5
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0